Instruction Format Ideas
If we only use 8 registers for one program, we don't need 4bits to specify registers, we only need 3bits. In that case we can use the extra bit for an immediate. And if we only use for example 8 operations in that program then we need only 3bits for operation instructions which would give us 2 more extra bits. That's a total of 3 bits for immediates.